big refactor, and fix (most) unit tests.
