+++ filternet/models/base_net.py
