+++ tests/datasets/test_opportunity.py
@@ -4,8 +4,8 @@ import os
 
 import pytest
 
-from datasets import opportunity as opp
-from datasets import sliding_window_x_y
+from filternet.datasets import opportunity as opp
+from filternet.datasets import sliding_window_x_y
 
 
 @pytest.fixture
