+++ tests/test_datasets.py
@@ -1,8 +1,8 @@
 # Copyright (C) 2020 Pet Insight  Project - All Rights Reserved
 
-import modeling_paper.datasets
+import filternet.datasets
 
 
 def test_datasets_dir():
-    assert "datasets" in modeling_paper.datasets.datasets_dir
-    print(modeling_paper.datasets.datasets_dir)
+    assert "datasets" in filternet.datasets.datasets_dir
+    print(filternet.datasets.datasets_dir)
