+++ tests/test_init.py
@@ -1,8 +1,8 @@
 # Copyright (C) 2020 Pet Insight  Project - All Rights Reserved
 
-import modeling_paper
+import filternet
 
 
 def test_base_dir():
-    assert "modeling_paper" in modeling_paper.base_dir
-    print(modeling_paper.base_dir)
+    assert "filternet" in filternet.base_dir
+    print(filternet.base_dir)
